Updated formatting and variable naming.
Updated indentation formatting to Sir Mirek of Addepar's specifications.
Appended this. to every variable name where it wouldn't cause an error...
FYI, the naming convention I've been using for XML layout file ids is to do
lowercase and underscore for the unique id part, then the item type on the
end in uppercase.  For example: a button to change your username would be
something like change_usernameButton.  If this irritates you we can change
it, but it helped me keep track of what was what (especially when changing
the contact selection from a button that linked to the contacts app to an
autocomplete textview that referenced a pre-loaded contacts list, which it
is now).